Separate DB into modules
